HumanZeek/requirements.txt
git+https://github.com/naya1503/pyrogram
TgCrypto
pytz
pytgcalls==3.0.0.dev24
yt-dlp
youtube-search-python
motor
dnspython==2.3.0
GitPython
feedparser
speedtest-cli
pykeyboard
bs4
Pillow==8.3.2
pymediainfo
opencv-python-headless
requests
ffmpeg-python
unidecode
attrify
pyromod==1.1.0
html2text
removebg
aiofiles
aiohttp
gpytranslate
gtts
wget
telegraph
heroku3
openai
httpx[http2]
uvloop
httpx
python-dotenv
py-extract
psutil
git+https://github.com/Itz-fork/fake-useragent.git
search_engine_parser
future
pytimeparse
flask
flask_restful